MCR-2707 move classses to better packages
