+++ mycore-base/src/test/java/org/mycore/datamodel/objectinfo/MCRObjectInfoEntityQueryResolverTest.java
@@ -34,6 +34,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mycore.backend.jpa.MCREntityManagerProvider;
 import org.mycore.backend.jpa.objectinfo.MCRObjectInfoEntity;
+import org.mycore.backend.jpa.objectinfo.MCRObjectInfoEntityQueryResolver;
 import org.mycore.common.MCRJPATestCase;
 import org.mycore.common.content.MCRURLContent;
 import org.mycore.common.xml.MCRXMLParserFactory;
@@ -44,8 +45,6 @@ import org.mycore.datamodel.classifications2.MCRCategory;
 import org.mycore.datamodel.classifications2.MCRCategoryID;
 import org.mycore.datamodel.classifications2.impl.MCRCategoryDAOImpl;
 import org.mycore.datamodel.classifications2.utils.MCRXMLTransformer;
-import org.mycore.datamodel.common.MCRObjectQuery;
-import org.mycore.datamodel.common.MCRObjectQueryResolver;
 import org.mycore.datamodel.metadata.MCRObjectID;
 import org.xml.sax.SAXParseException;
